+++ R/switch.R
@@ -39,11 +39,7 @@
 #'
 #' @export
 #'
-oanda_switch <- function() {
-
-  do_$switchServer()
-
-}
+oanda_switch <- function() do_$switchServer()
 
 #' ichimoku Internal Functions
 #'
@@ -59,10 +55,8 @@ do_ <- function() {
   livestore <- keystore <- instruments <- account <- NULL
 
   list(
-    getServer = function() {
-      server_type
-    },
-    switchServer = function() {
+    getServer = function() server_type,
+    switchServer = function()
       switch(server_type,
              practice = {
                server_type <<- "live"
@@ -73,8 +67,7 @@ do_ <- function() {
                server_type <<- "practice"
                livestore <<- keystore <<- instruments <<- account <<- NULL
                message("Default OANDA server switched to 'practice'")
-             })
-    },
+             }),
     getKey = function(server) {
       if (missing(server)) server <- server_type
       switch(server,
@@ -114,9 +107,15 @@ do_ <- function() {
         url <- switch(server,
                       practice = "https://api-fxpractice.oanda.com/v3/accounts",
                       live = "https://api-fxtrade.oanda.com/v3/accounts")
-        resp <- ncurl(url, headers = c("Authorization" = paste0("Bearer ", apikey),
-                                       "User-Agent" = .user_agent))
-        parsed <- parse_json(.subset2(resp, "data"))
+        resp <- ncurl(url, convert = FALSE, follow = TRUE,
+                      headers = c("Authorization" = paste0("Bearer ", apikey),
+                                  "User-Agent" = .user_agent))
+        .subset2(resp, "status") == 200L ||
+          stop("status code ", .subset2(resp, "status"), " - ",
+               fparse(.subset2(resp, "raw"), max_simplify_lvl = 3L, type_policy = 0L, int64_policy = 0L),
+               call. = FALSE)
+        parsed <- fparse(.subset2(resp, "raw"),
+                         max_simplify_lvl = 3L, type_policy = 0L, int64_policy = 0L)
         length(.subset2(parsed, "accounts")) || stop(parsed, call. = FALSE)
         account <<- parsed[["accounts"]][[1L]][["id"]]
       }
@@ -129,9 +128,15 @@ do_ <- function() {
         url <- paste0("https://api-fx", switch(server, practice = "practice", live = "trade"),
                       ".oanda.com/v3/accounts/", do_$getAccount(server = server, apikey = apikey),
                       "/instruments")
-        resp <- ncurl(url, headers = c("Authorization" = paste0("Bearer ", apikey),
-                                       "User-Agent" = .user_agent))
-        parsed <- parse_json(.subset2(resp, "data"))
+        resp <- ncurl(url, convert = FALSE, follow = TRUE,
+                      headers = c("Authorization" = paste0("Bearer ", apikey),
+                                  "User-Agent" = .user_agent))
+        .subset2(resp, "status") == 200L ||
+          stop("status code ", .subset2(resp, "status"), " - ",
+               fparse(.subset2(resp, "raw"), max_simplify_lvl = 3L, type_policy = 0L, int64_policy = 0L),
+               call. = FALSE)
+        parsed <- fparse(.subset2(resp, "raw"),
+                         max_simplify_lvl = 3L, type_policy = 0L, int64_policy = 0L)
         length(.subset2(parsed, "instruments")) || {
           warning(parsed,
                   "\nInstruments list could not be retrieved - falling back to internal data",