+++ R/utils.R
@@ -90,13 +90,11 @@ tradingDays <- function(x, holidays, ...) {
 #'
 grid_dup <- function(n, omit.id) {
   vec <- vector(mode = "list", length = n - 1L)
-  for (i in seq_along(vec)) {
+  for (i in seq_along(vec))
     vec[[i]] <- i * n + 1:i
-  }
   vec <- unlist(vec)
-  if (!missing(omit.id) && isTRUE(omit.id)) {
+  if (!missing(omit.id) && isTRUE(omit.id))
     vec <- c(vec, 1:n + n * (1:n - 1))
-  }
   vec
 }
 
@@ -141,7 +139,8 @@ xts_df <- function(x, keep.attrs) {
   `attributes<-`(df, c(list(names = c("index", dn2),
                             class = "data.frame",
                             row.names = .set_row_names(xlen)),
-                       if (!missing(keep.attrs) && isTRUE(keep.attrs)) .Call(ichimoku_look, x)))
+                       if (!missing(keep.attrs) && isTRUE(keep.attrs))
+                         .Call(ichimoku_look, x)))
 }
 
 #' Convert matrix to data.frame
@@ -177,9 +176,8 @@ matrix_df <- function(x, keep.attrs) {
   end <- 1:len * xlen
   attributes(x) <- NULL
   df <- vector(mode = "list", length = len)
-  for (i in seq_along(df)) {
+  for (i in seq_along(df))
     df[[i]] <- x[start[i]:end[i]]
-  }
   `attributes<-`(df, c(list(names = dn[[2L]],
                             class = "data.frame",
                             row.names = if (is.null(dn[[1L]])) .set_row_names(xlen) else dn[[1L]]),
@@ -224,7 +222,8 @@ df_merge <- function(...) {
                                 price = attr(dots[[1L]], "price"),
                                 timestamp = .Call(ichimoku_psxct, max(unlist(lapply(dots, attr, "timestamp")))),
                                 oanda = TRUE))
-    if (FALSE %in% .subset2(merge, "complete")) warning("Incomplete periods in merged dataframe - please check for possible duplicates", call. = FALSE)
+    if (FALSE %in% .subset2(merge, "complete"))
+      warning("Incomplete periods in merged dataframe - please check for possible duplicates", call. = FALSE)
   }
   merge
 }
@@ -267,9 +266,8 @@ df_append <- function(old, new, key = "time", keep.attr = "timestamp") {
   keep <- !.subset2(old, key) %in% .subset2(new, key)
   cnames <- attr(new, "names")
   df <- vector(mode = "list", length = length(new))
-  for (i in seq_along(df)) {
+  for (i in seq_along(df))
     df[[i]] <- c(.subset2(old, i)[keep], .subset2(new, i))
-  }
   `attributes<-`(df, `names<-`(
     list(cnames, "data.frame", .set_row_names(length(df[[1L]])), attr(new, keep.attr)),
     c("names", "class", "row.names", keep.attr)))
@@ -342,7 +340,8 @@ look <- function(x) {
 more <- function(rows) {
 
   is.ichimoku(lv <- .Last.value) || return(invisible())
-  print(lv, plot = FALSE, rows = if (missing(rows) || !is.numeric(rows)) attr(lv, "dim")[1L] else rows)
+  print(lv, plot = FALSE,
+        rows = if (missing(rows) || !is.numeric(rows)) attr(lv, "dim")[1L] else rows)
 
 }